+++ packages/whisker-svg/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Web SVG-DOM and Desktop raster Host implementations.
+
 ## [0.11.1](https://github.com/whiskerrs/whisker/compare/whisker-svg-v0.11.0...whisker-svg-v0.11.1) - 2026-08-12
 
 ### Other